minor name variants refactor
