+++ card/lib/card/name.rb
@@ -4,7 +4,14 @@ require_dependency "card/env"
 require "smart_name"
 
 class Card
+  # The SmartName class provides generalized of Card naming patterns
+  # (compound names, key-based variants, etc)
+  #
+  # Card::Name adds support for deeper card integration
   class Name < SmartName
+    include FieldsAndTraits
+    include Variants
+
     RELATIVE_REGEXP = /\b_(left|right|whole|self|user|main|\d+|L*R?)\b/
 
     self.params  = Card::Env # yuck!
@@ -33,52 +40,8 @@ class Card
       right && "*" == right[0, 1]
     end
 
-    def trait_name? *traitlist
-      junction? && begin
-        right_key = right_name.key
-        traitlist.find do |codename|
-          (card_id = Card::Codename[codename]) &&
-            (card = Card.quick_fetch card_id) &&
-            card.key == right_key
-        end.present?
-      end
-    end
-
-    def trait_name tag_code
-      (card_id = Card::Codename[tag_code]) &&
-        (card = Card.quick_fetch card_id) &&
-        [self, card.cardname].to_name
-    end
-
-    def trait tag_code
-      if (name = trait_name tag_code)
-        name.s
-      else
-        raise Card::Error::NotFound, "unknown codename: #{tag_code}"
-      end
-    end
-
-    def field tag_name
-      field_name(tag_name).s
-    end
-
     def code
       Card::Codename[Card.fetch_id self]
-    end
-
-    # returns full name for a field
-    def field_name tag_name
-      case tag_name
-      when Symbol
-        trait_name tag_name
-      else
-        tag_name = tag_name.to_s[1..-1] if tag_name.to_s[0] == "+"
-        [self, tag_name].to_name
-      end
-    end
-
-    def relative_field_name tag_name
-      field_name(tag_name).relative_name(self)
     end
 
     def relative_name context_name
@@ -96,14 +59,6 @@ class Card
         relative_name(context_name).key != absolute_name(context_name).key
       else
         s.match(/^\s*\+/)
-      end
-    end
-
-    def field_of? context_name
-      if context_name.present?
-        child_of?(context_name) && relative_name(context_name).length == 2
-      else
-        s.match(/^\s*\+[^+]+$/)
       end
     end
